Hack around Symbol.is_local() bug for global symbols
Unfortunately Symbol.is_local() in 3.7+ always return False for global
symbols when it should be returning True if we are in the top (module)
scope and the symbol is bound.

It's a bit ugly accessing Symbol's private attributes, but unfortunately
there doesn't seem to be another way of working around this. Hopefully
bug fixes will be rolled out soon and we can remove this hack.