Remove special handling for decorators
The grammar differs between Python version so avoid parsing it directly.
Instead make sure we can handle the deeper layers correctly. For modern
Python we get atoms, but for older versions we get dotted names. So we
also need to make sure we do proper obfuscation of the first part of
dotted names.